+++ tp2/2_nested_functions/main.hpp
@@ -1,8 +1,8 @@
 #ifndef MAIN_H
 #define MAIN_H
 
-int addGrade();
-int mean(int nbGrades);
+double addGrade();
+double mean(int nbGrades);
 int max(int a, int b);
 int max(int a, int b, int c);
 double getPriceWithTaxes(double prix);
